[feat] Extract hypervisor/src/riscv into library package arch/src/riscv
