+++ arch/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "arch"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
