+++ arch/src/riscv/mod.rs
@@ -1,3 +1,2 @@
 pub mod cpu;
 pub mod instruction;
-pub mod sbi;
